+++ package.json
@@ -10,10 +10,12 @@
     "@types/node": "^16.11.60",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
+    "@types/recoil": "^0.0.9",
     "@types/styled-components": "^5.1.26",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.5",
     "styled-components": "^5.3.6",
     "styled-normalize": "^8.0.7",
     "typescript": "^4.8.4",
