+++ src/state/userState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const userState = atom<string>({
+  key: 'MessangerPage/username',
+  default: '장영준',
+});
